Remove unnecessary "use_visual" and "multiagent" logic
We'll be sharing this environment with the community, and ease of
reading / understanding the wrapper may be useful when determining
how to best utilize it for training.

Since subclassing the v0.6 UnityEnv wrapper isn't a viable option
and we've deviated significantly from the basic gym wrapper, I've
removed the code from this repository which shouldn't be needed
for interacting with Obstacle Tower -- which is related to
`multiagent` and `use_visual`.